+++ router-tests/update-config-no-edg.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
+set -e
+
 # js composition
 
 ## using source code
 
 echo "Generating config using 'wgc router compose'"
-pnpx tsx --env-file ../cli/.env ../cli/src/index.ts router compose -i ../demo/graph-no-edg.yaml -o ../router-tests/testenv/testdata/config.json
+pnpx tsx ../cli/src/index.ts router compose -i ../demo/graph-no-edg.yaml -o ../router-tests/testenv/testdata/config.json
 
 ## using npm package