+++ studio/CHANGELOG.md
@@ -4,6 +4,12 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.162.0](https://github.com/wundergraph/cosmo/compare/studio@0.161.9...studio@0.162.0) (2026-03-18)
+
+### Features
+
+* add subgraph check extension outcome to proposals screen ([#2661](https://github.com/wundergraph/cosmo/issues/2661)) ([638af2f](https://github.com/wundergraph/cosmo/commit/638af2fd85cfcf715eb887b3f0606a898b2f7999)) (@wilsonrivera)
+
 ## [0.161.9](https://github.com/wundergraph/cosmo/compare/studio@0.161.8...studio@0.161.9) (2026-03-17)
 
 **Note:** Version bump only for package studio